+++ moderngl/scope.py
@@ -1,4 +1,10 @@
+import logging
+
+from moderngl.mgl import InvalidObject  # type: ignore
+
 __all__ = ['Scope']
+
+LOG = logging.getLogger(__name__)
 
 
 class Scope:
@@ -19,11 +25,17 @@ class Scope:
         - Restore the framebuffer.
     '''
 
-    __slots__ = ['mglo', 'ctx', 'extra']
+    __slots__ = ['mglo', 'ctx', '_framebuffer', '_textures', '_uniform_buffers', '_storage_buffers', '_samplers', 'extra']
 
     def __init__(self):
         self.mglo = None  #: Internal representation for debug purposes only.
         self.ctx = None  #: The context this object belongs to
+        # Keep references to keep this objects alive
+        self._framebuffer = None
+        self._textures = None
+        self._uniform_buffers = None
+        self._storage_buffers = None
+        self._samplers = None
         self.extra = None  #: Any - Attribute for storing user defined objects
         raise TypeError()
 
@@ -39,3 +51,18 @@ class Scope:
 
     def __exit__(self, *args):
         self.mglo.end()
+
+    def __del__(self):
+        LOG.debug(f"{self.__class__.__name__}.__del__ {self}")
+        if hasattr(self, "ctx") and self.ctx.gc_mode == "auto":
+            self.release()
+
+    def release(self):
+        LOG.debug(f"{self.__class__.__name__}.release() {self}")
+        if not isinstance(self.mglo, InvalidObject):
+            self._framebuffer = None
+            self._textures = None
+            self._uniform_buffers = None
+            self._storage_buffers = None
+            self._samplers = None
+            self.mglo.release()