+++ moderngl/src/Buffer.cpp
@@ -664,11 +664,10 @@ void MGLBuffer_Invalidate(MGLBuffer * buffer) {
 		return;
 	}
 
-	// TODO: decref
-
 	const GLMethods & gl = buffer->context->gl;
 	gl.DeleteBuffers(1, (GLuint *)&buffer->buffer_obj);
 
 	Py_TYPE(buffer) = &MGLInvalidObject_Type;
+	Py_DECREF(buffer->context);
 	Py_DECREF(buffer);
 }